requirements.txt
flask
pillow
diffusers
transformers
torch
accelerate
opencv-python
opencv-contrib-python
